+++ sympy/physics/mechanics/lagrange.py
@@ -4,6 +4,7 @@ __all__ = ['LagrangesMethod']
 
 from sympy import diff, zeros, Matrix, eye, sympify, Symbol
 from sympy.physics.vector import (dynamicsymbols, ReferenceFrame, Point)
+from sympy.physics.mechanics.functions import mat_inv_mul
 
 
 class LagrangesMethod(object):
@@ -316,30 +317,6 @@ class LagrangesMethod(object):
         else:
             return (Matrix(self._qdots)).col_join(self.forcing)
 
-    def _mat_inv_mul(self, A, B):
-        """Internal Function
-
-        Computes A^-1 * B symbolically w/ substitution, where B is not
-        necessarily a vector, but can be a matrix.
-
-        """
-
-        r1, c1 = A.shape
-        r2, c2 = B.shape
-        temp1 = Matrix(r1, c1, lambda i, j: Symbol('x' + str(j) + str(r1 * i)))
-        temp2 = Matrix(r2, c2, lambda i, j: Symbol('y' + str(j) + str(r2 * i)))
-        for i in range(len(temp1)):
-            if A[i] == 0:
-                temp1[i] = 0
-        for i in range(len(temp2)):
-            if B[i] == 0:
-                temp2[i] = 0
-        temp3 = []
-        for i in range(c2):
-            temp3.append(temp1.LDLsolve(temp2[:, i]))
-        temp3 = Matrix([i.T for i in temp3]).T
-        return temp3.subs(dict(list(zip(temp1, A)))).subs(dict(list(zip(temp2, B))))
-
     def rhs(self, inv_method=None):
         """ Returns equations that can be solved numerically
 
@@ -347,12 +324,14 @@ class LagrangesMethod(object):
         ==========
 
         inv_method : str
-            The specific sympy inverse matrix calculation method to use.
+            The specific sympy inverse matrix calculation method to use. For a
+            list of valid methods, see :py:method:
+            `~sympy.matrices.matrices.MatrixBase.inv`
 
         """
 
         if inv_method is None:
-            self._rhs = self._mat_inv_mul(self.mass_matrix_full,
+            self._rhs = mat_inv_mul(self.mass_matrix_full,
                                           self.forcing_full)
         else:
             self._rhs = (self.mass_matrix_full.inv(inv_method,
